requirements.txt
Flask==0.12.2
py-cpuinfo==3.3.0
psutil==5.3.1